Отлажены запросы на добавление, обновление и показ всех задач

+++ src/main/java/todolist/controller/AdminController.java
@@ -5,11 +5,16 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 import todolist.entity.Todo;
+import todolist.exceptions.AppError;
 import todolist.repository.TodoRepository;
+import todolist.services.AdminService;
 
+import java.security.Principal;
 import java.util.List;
 import java.util.Optional;
 
@@ -19,18 +24,23 @@ import java.util.Optional;
 public class AdminController {
 
     private final TodoRepository todoRepository;
+    private final AdminService adminService;
 
-    public AdminController(TodoRepository todoRepository) {
+    public AdminController(TodoRepository todoRepository, AdminService adminService) {
         this.todoRepository = todoRepository;
+        this.adminService = adminService;
     }
+
+    public Principal principal;
 
     @PostMapping(path = "/create")
     @ResponseStatus(HttpStatus.CREATED)
     public Todo createTodo(@RequestBody Todo todo) {
+        todo = adminService.createTodo(todo);
         return todoRepository.save(todo);
     }
 
-    @GetMapping(path = "/{id}")
+    @GetMapping(path = "/{id}", consumes = "application/json")
     public ResponseEntity<Todo> todoById(@PathVariable("id") Long id) {
         Optional<Todo> optional = todoRepository.findById(id);
         return optional.map(todo ->
@@ -38,7 +48,7 @@ public class AdminController {
                 new ResponseEntity<>(HttpStatus.NOT_FOUND));
     }
 
-    @GetMapping(path = "/getall")
+    @GetMapping(path = "/getall", consumes = "application/json")
     public ResponseEntity<List<Todo>> getToDos(@RequestHeader HttpHeaders headers) {
         return new ResponseEntity<>
                 (this.todoRepository.findAll(),
@@ -55,7 +65,8 @@ public class AdminController {
         return todoRepository.save(todo);
     }
 
-    @PatchMapping(path = "/path/{id}", consumes = "application/json")
+    // Обновление Заголовка
+    @PatchMapping(path = "/update/{id}", consumes = "application/json")
     public Todo pathTodo(
             @PathVariable("id") Long id,
             @RequestBody Todo patch
@@ -65,20 +76,28 @@ public class AdminController {
             todo.setTitle(patch.getTitle());
         }
 
-        if (todo.getDescription() != null) {
-            todo.setDescription(patch.getDescription());
-        }
         return todoRepository.save(todo);
     }
 
-    @DeleteMapping("/delete/{id}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteTodo(@PathVariable("id") Long id) {
-        try {
+//    @DeleteMapping("/delete/{id}")
+//    @ResponseStatus(HttpStatus.NO_CONTENT)
+//    public void deleteTodo(@PathVariable("id") Long id) {
+//        try {
+//            todoRepository.deleteById(id);
+//         } catch (Exception e) {
+//            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+//        }
+//    }
+
+
+    @RequestMapping(value = "/delete/{id}", consumes = "application/json", method = RequestMethod.DELETE)
+    public @ResponseBody String deleteTodo(@PathVariable("id") Long id) {
+
+        if (todoRepository.existsById(id)){
             todoRepository.deleteById(id);
-        } catch (Exception e) {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+            return "Задача удалена";
+        } else {
+            return "Такой задачи нет";
         }
     }
-
 }
